Backend/requirements.txt

Django>=4.2,<5.0

djangorestframework>=3.14,<4.0

djangorestframework-simplejwt>=5.2,<6.0

psycopg2-binary>=2.9,<3.0  

corsheaders>=3.13,<4.0

pytz>=2021

requests>=2.28

numpy>=1.24

pandas>=1.5

python-decouple>=3.6

PyJWT>=2.7

pydantic>=1.10

pandas 
numpy 
scikit-learn 
joblib 
tensorflow 
djangorestframework
joblib
tensorflow

# For frontend dev in React
node
vite
react
react-dom
zustand

# Tailwind CSS
postcss
postcss-cli
autoprefixer